Move more data definitions in sysFeatures from headers to bodies
