+++ src/sysFeatures_intel.c
@@ -104,70 +104,36 @@ int intel_cpu_msr_register_setter(const LikwidDevice_t device, uint32_t reg, int
     return HPMwrite(device->id.simple.id, MSR_DEV, reg, msrData);
 }
 
-
-
-/* Intel Turbo */
-int intel_cpu_turbo_test(void)
-{
-    uint32_t eax = 0x01, ebx, ecx = 0x0, edx;
-    CPUID(eax, ebx, ecx, edx);
-    if (field32(ecx, 7, 1) == 0)
-    {
-        DEBUG_PRINT(DEBUGLEV_DEVELOP, Intel SpeedStep not supported by architecture);
-        return 0;
-    }
-
-    return likwid_sysft_foreach_hwt_testmsr(MSR_IA32_MISC_ENABLE);
-}
-
-int intel_cpu_turbo_getter(const LikwidDevice_t device, char** value)
-{
-    if (intel_cpu_turbo_test())
-    {
-        return intel_cpu_msr_register_getter(device, MSR_IA32_MISC_ENABLE, 36, 1, true, value);
-    }
-    return -ENOTSUP;
-}
-
-int intel_cpu_turbo_setter(const LikwidDevice_t device, const char* value)
-{
-    if (intel_cpu_turbo_test())
-    {
-        return intel_cpu_msr_register_setter(device, MSR_IA32_MISC_ENABLE, 36, 1, true, value);
-    }
-    return -ENOTSUP;
-}
-
 static const _SysFeatureList* intel_arch_feature_inputs[] = {
-    &intel_cpu_prefetch_feature_list,
-    &intel_cpu_ida_feature_list,
-    &intel_cpu_turbo_feature_list,
-    &intel_uncorefreq_feature_list,
+    &likwid_sysft_intel_cpu_prefetch_feature_list,
+    &likwid_sysft_intel_cpu_ida_feature_list,
+    &likwid_sysft_intel_cpu_turbo_feature_list,
+    &likwid_sysft_intel_uncorefreq_feature_list,
     &likwid_sysft_intel_cpu_spec_ctrl_feature_list,
     NULL,
 };
 
 static const _SysFeatureList* intel_8f_arch_feature_inputs[] = {
-    &intel_cpu_prefetch_feature_list,
-    &intel_8f_cpu_feature_list,
-    &intel_cpu_ida_feature_list,
-    &intel_cpu_turbo_feature_list,
+    &likwid_sysft_intel_cpu_prefetch_feature_list,
+    &likwid_sysft_intel_8f_cpu_feature_list,
+    &likwid_sysft_intel_cpu_ida_feature_list,
+    &likwid_sysft_intel_cpu_turbo_feature_list,
     &likwid_sysft_intel_cpu_spec_ctrl_feature_list,
     NULL,
 };
 
 static const _SysFeatureList* intel_knl_arch_feature_inputs[] = {
-    &intel_knl_cpu_feature_list,
-    &intel_cpu_ida_feature_list,
-    &intel_cpu_turbo_feature_list,
+    &likwid_sysft_intel_knl_cpu_feature_list,
+    &likwid_sysft_intel_cpu_ida_feature_list,
+    &likwid_sysft_intel_cpu_turbo_feature_list,
     &likwid_sysft_intel_cpu_spec_ctrl_feature_list,
     NULL,
 };
 
 static const _SysFeatureList* intel_core2_arch_feature_inputs[] = {
-    &intel_core2_cpu_feature_list,
-    &intel_cpu_ida_feature_list,
-    &intel_cpu_turbo_feature_list,
+    &likwid_sysft_intel_core2_cpu_feature_list,
+    &likwid_sysft_intel_cpu_ida_feature_list,
+    &likwid_sysft_intel_cpu_turbo_feature_list,
     &likwid_sysft_intel_cpu_spec_ctrl_feature_list,
     NULL,
 };
