+++ src/sysFeatures_intel_prefetcher.c
@@ -61,6 +61,18 @@ int intel_cpu_l1_dcu_ip_setter(const LikwidDevice_t device, const char* value)
     return intel_cpu_msr_register_setter(device, MSR_PREFETCH_ENABLE, 3, 1, true, value);
 }
 
+static _SysFeature intel_cpu_prefetch_features[] = {
+    {"l2_hwpf", "prefetch", "L2 Hardware Prefetcher", intel_cpu_l2_hwpf_getter, intel_cpu_l2_hwpf_setter, DEVICE_TYPE_HWTHREAD, intel_cpu_l2_hwpf_register_test},
+    {"l2_adj_pf", "prefetch", "L2 Adjacent Cache Line Prefetcher", intel_cpu_l2_adj_pf_getter, intel_cpu_l2_adj_pf_setter, DEVICE_TYPE_HWTHREAD, intel_cpu_l2_hwpf_register_test},
+    {"l1_dcu", "prefetch", "DCU Hardware Prefetcher", intel_cpu_l1_dcu_getter, intel_cpu_l1_dcu_setter, DEVICE_TYPE_HWTHREAD, intel_cpu_l2_hwpf_register_test},
+    {"l1_dcu_ip", "prefetch", "DCU IP Prefetcher", intel_cpu_l1_dcu_ip_getter, intel_cpu_l1_dcu_ip_setter, DEVICE_TYPE_HWTHREAD, intel_cpu_l2_hwpf_register_test},
+    //{"data_pf", "Data Dependent Prefetcher", DEVICE_TYPE_HWTHREAD},
+};
+
+const _SysFeatureList likwid_sysft_intel_cpu_prefetch_feature_list = {
+    .num_features = ARRAY_COUNT(intel_cpu_prefetch_features),
+    .features = intel_cpu_prefetch_features,
+};
 
 /*********************************************************************************************************************/
 /*                          Intel 0x8F prefetchers                                                                   */
@@ -76,6 +88,14 @@ int intel_cpu_l2_multipath_pf_setter(const LikwidDevice_t device, const char* va
     return intel_cpu_msr_register_setter(device, MSR_PREFETCH_ENABLE, 6, 1, true, value);
 }
 
+static _SysFeature intel_8f_cpu_features[] = {
+    {"l2_multipath_pf", "prefetch", "L2 Adaptive Multipath Probability Prefetcher", intel_cpu_l2_multipath_pf_getter, intel_cpu_l2_multipath_pf_setter, DEVICE_TYPE_HWTHREAD, intel_cpu_l2_hwpf_register_test}
+};
+
+const _SysFeatureList likwid_sysft_intel_8f_cpu_feature_list = {
+    .num_features = ARRAY_COUNT(intel_8f_cpu_features),
+    .features = intel_8f_cpu_features,
+};
 
 /*********************************************************************************************************************/
 /*                          Intel Knights Landing prefetchers                                                        */
@@ -100,6 +120,15 @@ int intel_knl_l2_hwpf_setter(const LikwidDevice_t device, const char* value)
     return intel_cpu_msr_register_setter(device, MSR_PREFETCH_ENABLE, 1, 1, true, value);
 }
 
+static _SysFeature intel_knl_cpu_prefetch_features[] = {
+    {"l2_hwpf", "prefetch", "L2 Hardware Prefetcher", intel_knl_l2_hwpf_getter, intel_knl_l2_hwpf_setter, DEVICE_TYPE_HWTHREAD, intel_cpu_l2_hwpf_register_test},
+    {"l1_dcu", "prefetch", "DCU Hardware Prefetcher", intel_knl_l1_dcu_getter, intel_knl_l1_dcu_setter, DEVICE_TYPE_HWTHREAD, intel_cpu_l2_hwpf_register_test},
+};
+
+const _SysFeatureList likwid_sysft_intel_knl_cpu_feature_list = {
+    .num_features = ARRAY_COUNT(intel_knl_cpu_prefetch_features),
+    .features = intel_knl_cpu_prefetch_features,
+};
 
 /*********************************************************************************************************************/
 /*                          Intel Core2 prefetchers                                                                  */
@@ -150,6 +179,18 @@ int intel_core2_l1_dcu_ip_setter(const LikwidDevice_t device, const char* value)
     return intel_cpu_msr_register_setter(device, MSR_IA32_MISC_ENABLE, 39, 1, true, value);
 }
 
+static _SysFeature intel_core2_cpu_prefetch_features[] = {
+    {"hwpf", "prefetch", "Hardware prefetcher operation on streams of data", intel_core2_l2_hwpf_getter, intel_knl_l2_hwpf_setter, DEVICE_TYPE_HWTHREAD, intel_core2_l2_hwpf_register_test},
+    {"adj_pf", "prefetch", "Adjacent Cache Line Prefetcher", intel_core2_l2_adjpf_getter, intel_core2_l2_adjpf_setter, DEVICE_TYPE_HWTHREAD, intel_core2_l2_hwpf_register_test},
+    {"l1_dcu", "prefetch", "DCU L1 data cache prefetcher", intel_core2_l1_dcu_getter, intel_core2_l1_dcu_setter, DEVICE_TYPE_HWTHREAD, intel_core2_l2_hwpf_register_test},
+    {"l1_dcu_ip", "prefetch", "DCU IP Prefetcher", intel_core2_l1_dcu_ip_getter, intel_core2_l1_dcu_ip_setter, DEVICE_TYPE_HWTHREAD, intel_core2_l2_hwpf_register_test},
+};
+
+const _SysFeatureList likwid_sysft_intel_core2_cpu_feature_list = {
+    .num_features = ARRAY_COUNT(intel_core2_cpu_prefetch_features),
+    .features = intel_core2_cpu_prefetch_features,
+};
+
 /*********************************************************************************************************************/
 /*                          Intel Dynamic Acceleration                                                               */
 /*********************************************************************************************************************/
@@ -180,3 +221,14 @@ int intel_core2_ida_getter(const LikwidDevice_t device, char** value)
 int intel_core2_ida_setter(const LikwidDevice_t device, const char* value) {
     return intel_cpu_msr_register_setter(device, MSR_IA32_MISC_ENABLE, 38, 1, true, value);
 }
+
+static _SysFeature intel_cpu_ida_features[] = {
+    // TODO does this really belong into the "prefetcher" category?
+    {"ida", "prefetch", "Intel Dynamic Acceleration", intel_core2_ida_getter, intel_core2_ida_setter, DEVICE_TYPE_HWTHREAD, intel_core2_ida_tester},
+};
+
+const _SysFeatureList likwid_sysft_intel_cpu_ida_feature_list = {
+    .num_features = ARRAY_COUNT(intel_cpu_ida_features),
+    .tester = intel_core2_ida_tester,
+    .features = intel_cpu_ida_features,
+};
